Add email setup guide, booking redirect routes, and verification service
- Created EMAIL_SETUP.md for configuring email functionality, including Gmail, Outlook, Yahoo, and custom SMTP settings.
- Implemented booking redirect routes in booking.py to handle flight booking redirects via SerpAPI, including error handling and HTML response generation.
- Developed booking_redirect_service.py to manage fetching booking options from SerpAPI and generating redirect HTML.
- Added verification_service.py to handle email verification codes for user registration, including code generation, storage, and email sending.

+++ app/routes/recommendations.py
@@ -33,8 +33,8 @@ class FlightSelectionRequest(BaseModel):
     arrival_city: str
     departure_time: datetime
     airline: str
-    price: int
-    overall_score: int
+    price: float
+    overall_score: float
     cabin_class: str = "economy"
